Tareas en AccountController

+++ src/main/java/com/mikadifo/controllers/AccountController.java
@@ -3,13 +3,17 @@ package com.mikadifo.controllers;
 import com.mikadifo.models.db_tables.City;
 import com.mikadifo.models.db_tables.User;
 import com.mikadifo.models.table_statements.UserDB;
+import static java.lang.Character.*;
 import java.net.URL;
+import java.util.Optional;
 import java.util.ResourceBundle;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.TextField;
 import javafx.scene.input.KeyEvent;
@@ -53,15 +57,34 @@ public class AccountController implements Initializable {
     public void init(Scene scene, UserDB user) {
         btnUpdate.getScene().getStylesheets().add("/styles/account.css");
         currentUser = user;
+        
     }
 
     @FXML
     private void onUsernameKeyReleased(KeyEvent event) {
+        String characterTyped = event.getCharacter();
+        if (!characterTyped.isEmpty()) {
+            char val = characterTyped.charAt(0);
+            
+            if (!(isLetterOrDigit(val)|| val == '_' )|| txtUsername.getText().length() > 49) {
+                event.consume();
+            }
+        }
+        
     }
 
 
     @FXML
     private void onLoginKeyReleased(KeyEvent event) {
+        String characterTyped = event.getCharacter();
+
+        if (!characterTyped.isEmpty()) {
+            char val = characterTyped.charAt(0);
+
+            if (!isDigit(val) || txtLogin.getText().length() > 9) {
+                event.consume();
+            }
+        }
     }
 
     @FXML
@@ -72,12 +95,18 @@ public class AccountController implements Initializable {
 
     @FXML
     private void onDeleteAccountClick(ActionEvent event) {
-        getUserFromView().delete();
+        boolean isOk = showAlert(Alert.AlertType.CONFIRMATION, null, "¿Esta seguro que desea eliminar la cuenta?");
+
+        if (isOk) getUserFromView().delete();
+        
     }
 
     @FXML
     private void onUptadeClick(ActionEvent event) {
-        getUserFromView().update();
+        boolean isOk = showAlert(Alert.AlertType.CONFIRMATION, null, "¿Esta seguro que desea actualizar su cuenta?");
+        
+        if (isOk) getUserFromView().update();
+        
     }
 
     private UserDB getUserFromView() {
@@ -94,5 +123,15 @@ public class AccountController implements Initializable {
 
     @FXML
     private void onCityKeyReleased(KeyEvent event) {
+    }
+    
+    private boolean showAlert(Alert.AlertType alertType, String header, String message) {
+	Alert alert = new Alert(alertType);
+
+        alert.setHeaderText(header);
+        alert.setTitle(null);
+        alert.setContentText(message);
+
+	return alert.showAndWait().get() == ButtonType.OK;
     }
 }
